+++ src/output_functions.cc
@@ -11,7 +11,7 @@ void output_edge_list(const edge_list_t & edge_list, std::ostream & stream)
 
 void output_adj_list(const adj_list_t & adj_list, std::ostream & stream)
 {
-  for (unsigned int n = 0; n < adj_list.size(); ++n)
+  for (size_t n = 0; n < adj_list.size(); ++n)
   {
     stream << n << " : ";
     for (auto neighbour = adj_list[n].begin(); neighbour != adj_list[n].end(); ++neighbour)
